==> analysis/04_Biodiversity.Rmd <==
---
title: "Biodiversity Analysis"
author: "Sola Takahashi"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/04_Biodiversity/")
```

# Setting the Environment

## Set the seed
```{r set-seed}
set.seed(158369)
```


## Load Libraries
```{r load-packages}
pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq, 
               install = FALSE)
```


## Load in Data
```{r load-data}
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Intuition Check
min(sample_sums(raw_preprocessed_physeq))
# output: 1836

# Make a metadata dataframe
metadata_df <- 
  raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame() 

#head(metadata_df)

# Setting colors for the stations
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")


```


# Goals

1. Calculate the Hill Diversity of the samples
2. Evaluate the rarefaction curves
3. Evaluate the Diversity values
4. Make notes of specific samples and their seq depth 

# Diversity Calculations with iNEXT

```{r calc-div}
# prepare input data
iNEXT_input_df <-
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()

# Quick check
dim(iNEXT_input_df)

# Run iNEXT: Calculate Hill Numbers
# Note that: Species in ROWS and samples in COLUMNS
# Remember to set the seed!
# iNEXT_data <- iNEXT(iNEXT_input_df, q = c(0,1,2), 
#                    datatype = "abundance")

# Save the file
#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the Diversity
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
typeof(iNEXT_data)
# size based estimator is the best for incidence data
# this type of list is less stringent data standards (S3), don't need anchoring of sample/column names (S4 - with phyloseq object)

```


# Plot Diversity
```{r plot-rarefaction}
# Prepare colors
color_df <-
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
  
# Check
head(color_df)

# Rename column
colnames(color_df)[1] <- "names"

# Make a helper dataframe for plotting with colors
# replace . and remove X
iNEXT_color_df <-
color_df %>%
  # fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"), 
         names = gsub(names, pattern = "X", replace = "")) %>%
  # merge iwth metadata
  left_join(metadata_df, by = "names") %>%
  # merge with colors for plotting with ggNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
  


```

# Plot Rarefaction with 'ggiNEXT'

```{r ggiNEXT, fig.width = 8, fig.height= 3.5}
# Plot rarefaction
# rarefaction v extrapolation curve, type = 1

# Order q: 
# 0 = RIchness/ total number taxa
# 1 = Exponential Shannon/ Number of common taxa
# 2 = Inverse simpson / Number of dominant taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") +
  facet_wrap(~Order.q, scales = "fixed") +
  scale_color_manual(values = iNEXT_color_df$station_colors, guide = FALSE) +
  scale_fill_manual(values = iNEXT_color_df$station_colors, guide = FALSE) +
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)), 
                     guide = FALSE) +
  theme(legend.position = "none")
# ggplot shapes, 17= triangle
# Analysis:
# in soils and water -> microbial seed bank of a bunch of dormant species 
# want to see our data plateaus so that we know we've sequenced everything
# Simpson and Shannon plateaus earlier because less sensitive because fewer numbers of ASVs that are abundant/common, so more likely to sample them earlier on in the sampling. Richness is very sensitive.

```

# Manually plot Diversity

## Rarefaction
```{r iNEXT-manual}

iNEXT_manual_df <- 
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # fix sample names
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X", replace = "")) %>%
  #head()
  # join with meta data
  left_join(., metadata_df, by = "names") %>%
  # Add colors to data frame
  left_join(., data.frame(station_colors = station_colors,
                          station = names(station_colors)),
            by = "station") 
# Inspect
dim(iNEXT_manual_df)
str(iNEXT_manual_df)

# Plot it - Rarefaction curve
iNEXT_manual_df %>%
  # Filter out rows that are calculated by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Make the actual plot
  ggplot(aes( x= m, y = qD, color = station, group = names)) +
  geom_line() +
  #geom_point() +
  # facet with the station
  facet_grid(Order.q~ station, scales = "fixed") + 
  scale_color_manual(values = station_colors) +
  theme(legend.position = "bottom")

```


# Environmental Variable Check
```{r environmental-pairs, fig.width= 8, fig.height=8}
#head(metadata_df)

# Pull out environmental variables
env_df <-
  metadata_df %>%
  dplyr::select(names, water_tempC:DO_mgL)
# Inspect
# head(env_df)

# Plot the correlations
pairs(dplyr::select(env_df, -names), upper.panel = NULL)
```


# Diversity vs Salinity
```{r div-vs-salinity, fig.height=3.5, fig.width=6}
# PSU = practical salinity unit 
# 1 PSU = 1g of salt per 1000 g of water
# 1 PSU = 1 ppt (parts per thousand)
# Freshwater = 0 - 0.5
# Brackish = 0.5 - 30
# Marine = >30

iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = salinity_psu, y = qD)) +
  facet_wrap(.~ Order.q, scales = "free") +
  geom_point(aes(color = station)) +
  stat_smooth(method = "lm", formula = y ~poly(x, 2)) + 
  labs(x = "Salinity (PSU)", y = "# of ASVs") +
  scale_color_manual(values = station_colors) +
  theme(legend.position = "bottom")
# more dominant taxa in freshwater and marine water than brackish water 
```

# Session Information
```{r session-info}
devtools::session_info()
```
